% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ChaudhuriChristofidesDatapij.R
\docType{data}
\name{ChaudhuriChristofidesDatapij}
\alias{ChaudhuriChristofidesDatapij}
\title{Matrix of the second-order inclusion probabilities}
\usage{
ChaudhuriChristofidesDatapij
}
\description{
This dataset consists of a square matrix of dimension 100 with the first and second order inclusion probabilities
for the units included in sample \eqn{s}, drawn from a population of size \eqn{N=417} according to a
sampling with unequal probabilities (probability proportional to agricultural subsidies in the previous year).
}
\examples{
data(ChaudhuriChristofidesDatapij)
#Now, let select only the first-order inclusion probabilities
diag(ChaudhuriChristofidesDatapij)
}
\seealso{
\code{\link{ChaudhuriChristofides}}

\code{\link{ChaudhuriChristofidesData}}
}
\keyword{datasets}

